Rework Steering for Separate Claude skill
Rework the steering structure so that we can separably update the
Claude Skill. This is commit 1/n towards a final form updated Claude
Skill. Includes bash scripts for quick reference, reworking of the
workflows to leverage MCP server use instead of a Kiro Power, an
mcp specific guidance reference doc, and refactoring of how references
are evaluated.

Co-authored-by: anwesham-lab <amukherjee973@gmail.com>