Add aurora-dsql-samples as submodule for code embedding (#1)
The goal is for dsql-starter-kit to be narrative markdown entry-point and
for dsql-samples to be the single source of truth for code which will be
referenced. We want to minimize the amount of human intervention needed
to keep them in sync.

This will allow us to embed code samples directly in documentation:
- Adds aurora-dsql-samples as git submodule at documentation/docs/samples/
- Enables single source of truth: all code lives in samples repo
- Allows inline code display in docs using MkDocs snippets (--8<-- syntax)

Follow-up changes will move code and create the references.

Co-authored-by: Aleksandar Maksimovic <alemaksi@amazon.com>